updated using cast

+++ src/main_window.cpp
@@ -1157,7 +1157,7 @@ QImage Mat2QImage(cv::Mat const& src)
 {
     cv::Mat temp;
     cvtColor(src, temp,cv::COLOR_BGR2RGB);
-    QImage dest((const uchar *) temp.data, temp.cols, temp.rows, temp.step, QImage::Format_RGB888);
+    QImage dest((const uchar *) temp.data, static_cast<int>(temp.cols), static_cast<int>(temp.rows), static_cast<int>(temp.step), QImage::Format_RGB888);
     dest.bits();
     return dest;
 }
